# Let Messaging#fetch_schema_by_id use the schemas_by_id cache as well to improve encoding performance when the schema ID is given